requirements.txt
fastapi
uvicorn
python-telegram-bot
psycopg2-binary